ai_agent/pyproject.toml
[project]
name = "ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "grpcio>=1.71.0",
    "grpcio-tools>=1.71.0",
    "mcp[cli]>=1.9.1",
    "openai>=1.82.0",
    "python-dotenv>=1.1.0",
]
